+++ CHANGELOG.md
@@ -5,6 +5,10 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- Authentication token support through `UPCLOUD_TOKEN` environment variable and `token` config parameter
+
 ## [1.6.0] - 2025-05-23
 
 ### Added